add key action intents
