+++ lib/searchfield.dart
@@ -1,1 +1,2 @@
+export 'src/decoration.dart';
 export 'src/searchfield.dart';
